pvrparse: check all values from video enumeration
This commit drops the default case and uses all available codecs as
suggested in OpenPLi/enigma2#1914 (comment).
Also it uses the UNKNOWN enum value as suggsted in OpenPLi/enigma2#1914
(comment)

So once new codecs introduce we are going to have a warning on logs.

It would be nice to silence / fix all remaining warnigns and enable
-Werror to force all warnings as errors.
So we can catch erlier un-noticed issues.